feat: Refactor CreateTodo component to accept createTodo prop
The CreateTodo component was modified to accept the createTodo prop, which is a function for creating a new todo. This change allows the component to pass the user input (title, description, and image) to the createTodo function when the user clicks the submit button. The createTodo function is defined in the useTodo hook and is responsible for making an API request to create the todo. The component also resets the input fields after a successful submission.

Note: Please remove any unnecessary details and meta information from the commit message before finalizing it.