feat: Add Page component for readme
A new Page component was added to the app/readme directory. This component renders a simple "page" div and can be used to display content on the readme page. This addition enhances the structure and organization of the readme page.

Note: Please remove any unnecessary details and meta information from the commit message before finalizing it.